pyproject.toml
[tool.poetry]
name = "ldap-auth"
version = "0.1.0"
description = "LDAP auth with JWT"
authors = ["Sergei Demchenko <hb13718@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-ldap = "^3.4.3"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
